[UPD] Refactored the architecture for simpler usage.           The Editor is now only used to set some generation settings, that the terrainManager use (but he isn't forced to load them from here).

+++ Assets/Scripts/Terrain/Terrain/SimpleTerrainManager.cs
@@ -59,7 +59,7 @@ namespace Terrain
                     {
                         string chunkName = string.Format("chunk({0},{1},{2})", chunkX, chunkY, chunkZ);
                         GameObject chunkGO = new GameObject(chunkName);
-                        Vector3 chunkPosition = new Vector3(chunkX * (int)terrainProperties.chunkSize.x, chunkY * (int)terrainProperties.chunkSize.y, chunkZ * (int)terrainProperties.chunkSize.z);
+                        Vector3 chunkPosition = new Vector3(chunkX * terrainProperties.chunkSize.x, chunkY * terrainProperties.chunkSize.y, chunkZ * terrainProperties.chunkSize.z);
                         chunkGO.transform.SetParent(terrainGO.transform);
                         chunkGO.transform.SetLocalPositionAndRotation(chunkPosition, Quaternion.identity);
                         // Add the MeshFilter and MeshRenderer components
